move dsl into test
